Refactor source code and use browserify to link all files after babel transplied source code.
