+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/viewer.js",
   "scripts": {
     "build": "npm run lint && npm run compile",
-    "compile": "babel src -d lib",
+    "compile": "babel src -d lib/src && browserify lib/src/viewer.js -o lib/viewer.js",
     "lint": "eslint src/*.js",
     "lint-test": "eslint test/test.js",
     "test-build": "bash test/build.sh",
@@ -26,6 +26,7 @@
   "devDependencies": {
     "babel-cli": "^6.11.4",
     "babel-preset-es2015": "^6.13.2",
+    "browserify": "^13.1.0",
     "chai": "^3.5.0",
     "eslint": "^3.2.2",
     "express": "^4.14.0",
