+++ src/config.js
@@ -1,0 +1,20 @@
+export const DefaultProjection = 'EPSG:4326';
+
+export const UrlSymbals = {
+  "SemiColon": '_-_',
+  "Colon": '___',
+  "Comma": '_'
+};
+
+export const maxOpacity = 1.0;
+export const minOpacity = 0.1;
+
+export const extentUpdateDelay = 200;
+
+export default {
+  DefaultProjection,
+  UrlSymbals,
+  maxOpacity,
+  minOpacity,
+  extentUpdateDelay
+};
