+++ src/generalHelpers.js
@@ -1,0 +1,12 @@
+/*eslint no-console: "off"*/
+export const log = console.log.bind(console);
+export const info = console.info.bind(console);
+export const warn = console.warn.bind(console);
+export const error = console.error.bind(console);
+
+export {
+  log,
+  info,
+  warn,
+  error
+};
